Add curl-cffi HTTP client
