+++ pyproject.toml
@@ -50,6 +50,7 @@ aioshutil = "^1.3"
 beautifulsoup4 = { version = "^4.12.3", optional = true }
 colorama = "^0.4.6"
 cookiecutter = "^2.6.0"
+curl-cffi = "^0.7.1"
 docutils = "^0.21.0"
 eval-type-backport = "^0.2.0"
 html5lib = { version = "^1.1", optional = true }
