+++ src/crawlee/basic_crawler/__init__.py
@@ -1,6 +1,5 @@
 from .basic_crawler import BasicCrawler, BasicCrawlerOptions
 from .context_pipeline import ContextPipeline
 from .router import Router
-from .types import BasicCrawlingContext
 
-__all__ = ['BasicCrawler', 'BasicCrawlerOptions', 'ContextPipeline', 'Router', 'BasicCrawlingContext']
+__all__ = ['BasicCrawler', 'BasicCrawlerOptions', 'ContextPipeline', 'Router']
