+++ src/crawlee/basic_crawler/context_pipeline.py
@@ -4,14 +4,14 @@ from typing import Any, AsyncGenerator, Awaitable, Callable, Generator, Generic,
 
 from typing_extensions import TypeVar
 
-from crawlee.basic_crawler.errors import (
+from crawlee.errors import (
     ContextPipelineFinalizationError,
     ContextPipelineInitializationError,
     ContextPipelineInterruptedError,
     RequestHandlerError,
     SessionError,
 )
-from crawlee.basic_crawler.types import BasicCrawlingContext
+from crawlee.types import BasicCrawlingContext
 
 TCrawlingContext = TypeVar('TCrawlingContext', bound=BasicCrawlingContext, default=BasicCrawlingContext)
 TMiddlewareCrawlingContext = TypeVar('TMiddlewareCrawlingContext', bound=BasicCrawlingContext)
