+++ src/crawlee/http_clients/__init__.py
@@ -1,4 +1,5 @@
 from .base_http_client import BaseHttpClient, HttpCrawlingResult, HttpResponse
-from .httpx_client import HttpxClient
+from .curl_cffi_http_client import CurlCffiHttpClient
+from .httpx_http_client import HttpxHttpClient
 
-__all__ = ['BaseHttpClient', 'HttpCrawlingResult', 'HttpResponse', 'HttpxClient']
+__all__ = ['BaseHttpClient', 'CurlCffiHttpClient', 'HttpCrawlingResult', 'HttpResponse', 'HttpxHttpClient']
