+++ src/crawlee/http_crawler/types.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from crawlee.basic_crawler.types import BasicCrawlingContext
 from crawlee.http_clients import HttpCrawlingResult
+from crawlee.types import BasicCrawlingContext
 
 
 @dataclass(frozen=True)
