+++ tests/unit/cli/test_create.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import os
 from unittest.mock import Mock
