+++ tests/unit/statistics/test_persistence.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from crawlee.statistics import Statistics
