+++ tests/unit/test_configuration.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from crawlee.configuration import Configuration
